feat: add DisplayName to graphDesc (#180)
Decouples displayName from ID that graph is stored under. IDs are assigned when adding a graph to the engine.

Most visible effect of this is that all graphs are displayed with a capital letter so they are sorted properly on the web page. Later on grpah IDs might be changed to something that's independent of the graph Name (UUID or numeric ID). This will also simplify renaming of graphs which isn't supported at the moment.