fix(fritz): create only one alert if device with multiple functions goes offline (#249)
When the FritzBox looses contact to a device with multiple function (like a remote control), it marks all of the entities "not present". This raises multiple alerts.
Newer APIs give additional information for these "sub units". This additional object does not appear for the main device, so we do not raise a not-present-alert if the object exists.